requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.5.1
cbor==1.0.0
cchardet==2.1.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cytoolz==0.11.2
dataclasses==0.6
elastic-transport==8.1.2
elasticsearch==8.2.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fake-useragent==0.1.11
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.6.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.4
pandas==1.4.2
parsimonious==0.8.1
protobuf==3.20.1
pycares==4.1.2
pycparser==2.21
pycryptodome==3.14.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.3
pytz==2022.1
websockets==9.1
yarl==1.7.2
